The login model has been adjusted to utilize multiple providers at the same time. The login type that is first used to authenticate the user is referred to as the 'primary' login. Users may later authorize additional providers (i.e. facebook, twitter, google+?) to provide additional features.  The 'primary' login type does not change.
Code paths for Facebook posting are implemented.

The Facebook login process will be started when the user clicks the
Facebook Post button, if there is no existing Facebook login
The Facebook post will be made immediately upon pressing the Facebook Post
button, if Facebook permissions are already available.  Otherwise, the
post is made asynchronously when the Facebook API handler returns.  This
is because the UI thread may not wait until authentication has returned
before executing the post.

The serialized login data object has changed, users will have to login /
authenticate once again.

Used LollapaloozerApplication object to provide functionality for showing
an error dialog from any point in the application, by always knowing which
Activity is in the foreground.
The LollapaloozerApplication object will help to detect if multiple
instances of an Activity are being created by keeping track of
LollapaloozerActivity and ChooseLoginActivity objects created

The onActivityResult() value of 'request' parameter used by the Facebook
API has been identified as 32665 or "FBOOK" on a telephone keypad.  Code
has been updated so that unexpected onActivityResult() calls are not
handled as a response from the Facebook API.
The Facebook API requires the application to implement onActivityResult()
but does not clearly specify which value the API uses for the 'request'
field.  Since multiple sources can call onActivityResult(), in the past
this required handling a response from Facebook as the default case.

Identified the following bugs:
The application can sometimes be made to crash by clicking on a set with
existing rating/comment and then clicking the FB button as quickly as
possible
The application often fetches set data twice in a row
When logging in to an account with existing comments and ratings, the
ratings are not properly displayed until ratings are submitted or any
day's info is requested again