Twitter tweeting has been implemented The Twitter application 'Lollapaloozer' registered at dev.twitter.com has been changed from Read Only to Read/Write Access
Buttons to choose your primary login type have had images added

Buttons to post your rating to Facebook or Twitter have been enhanced with
images in the following ways:
If the user has not authenticated with the social network provider, a
small logo image for that provider is shown on a small button.
If the user has authenticated with the social network provider, a larger
"Share" image is shown on a larger size button
If the user has authenticated with both Facebook and Twitter, the regular
'Rate It' button appears to* move up in order to make room.
*Actually implemented with two buttons, where one button is set to visible
and the other is not visible.