Added log category SET_TIME_OPERATIONS so operations involving stored Set date/time and current date/time are easily debugged
Added method in CalendarUtils to return the current time in 24-hour format
as an integer between 0 and 2359

Behavior implemented that if the user clicks on a set that already
happened, the login or rate dialog is presented as usual.  However if the
user clicks a set that is in the future, the Alert dialog is presented.
The change in workflow also removes the ability to rate sets in the
future.

Further development of alert dialog box:

(The alert dialog box is only shown when the user clicks a set that has
not happened yet)

If there is no alert saved for this set, the user is offered the following
choice:

A) Alert me (x) minutes before the set begins  (X is a number input field,
default currently 15)

B) Never mind (same as cancelling the dialog

If there is already an alert saved for this set, the user is offered the
following choice

A) Alert me (x) minutes before the set begins (allowing the user to change
their previous setting
B) Cancel this alert